updates to build

+++ build.rb
@@ -27,16 +27,16 @@ Dir.chdir(rubygem_dir) do
   `git checkout iso_build_command`
   `bundle install`
 
-isos = `bundle exec ruby -I lib/ exe/simp-metadata release -v #{version} -w simp-metadata,https://github.com/brandonrdn/simp-metadata isos #{platform}`.split("\n")
+  isos = `bundle exec ruby -I lib/ exe/simp-metadata release -v #{version} -w simp-metadata,https://github.com/brandonrdn/simp-metadata isos #{platform}`.split("\n")
 
 # Copy Base ISOs to build directory
-isos.each do |iso|
-  next unless iso.include?(distribution)
-  puts "=== Copying #{iso} ==="
-  FileUtils.copy "/data/community-download/simp/ISO/base_isos/#{iso}", "#{iso_cache}" if File.exist?("/data/community-download/simp/ISO/base_isos/#{iso}")
-end
+  isos.each do |iso|
+    next unless iso.include?(distribution)
+    puts "=== Copying #{iso} ==="
+    FileUtils.copy "/data/community-download/simp/ISO/base_isos/#{iso}", "#{iso_cache}" if File.exist?("/data/community-download/simp/ISO/base_isos/#{iso}")
+  end
 
-heredoc = <<-HERDOC
+  heredoc = <<-HERDOC
 bundle exec ruby -I lib/
 exe/simp-metadata
 build iso
@@ -45,34 +45,37 @@ build iso
 --iso_cache #{iso_cache}
 --preserve
 -w simp-metadata,https://github.com/brandonrdn/simp-metadata
-HERDOC
+  HERDOC
 
-build_command = heredoc.tr("\n", ' ')
+  build_command = heredoc.tr("\n", ' ')
 
-system("#{build_command}")
+  system("#{build_command}")
 
 # Move files to binaries dir
 
 # Create binaries folders
-FileUtils.makedirs("#{binaries_dir}/{ISO,Tarballs,RPMs}")
+  FileUtils.makedirs("#{binaries_dir}/{ISO,Tarballs,RPMs}")
 
 # Move platform specific information
-platforms = `bundle exec ruby -I lib/ exe/simp-metadata release -v #{version} -w simp-metadata,https://github.com/brandonrdn/simp-metadata platforms`.split("\n")
+  platforms = `bundle exec ruby -I lib/ exe/simp-metadata release -v #{version} -w simp-metadata,https://github.com/brandonrdn/simp-metadata platforms`.split("\n")
 
-platforms.each do |dir|
-  if File.exist?("#{currentdir}/#{dir}")
-    Dir.chdir("#{currentdir}/#{dir}") do
-      # Copy ISO
-      iso_file = Dir.glob(File.join("**", 'SIMP', "*.iso"))
-      File.move iso_file, "#{binaries_dir}/ISO"
+  platforms.each do |dir|
+    next unless dir.include?(platform)
+    if File.exist?("#{currentdir}/#{dir}")
+      Dir.chdir("#{currentdir}/#{dir}") do
+        # Copy ISO
+        iso_file = Dir.glob(File.join("**", 'SIMP', "*.iso"))
+        File.move iso_file, "#{binaries_dir}/ISO"
 
-      # Copy Overlay Tarball
-      tar_files = Dir.glob(File.join("**", "SIMP", "*.tar.gz"))
-      tar_files.each {|file| File.move file, "#{binaries_dir}/Tarballs"}
+        # Copy Overlay Tarball
+        tar_files = Dir.glob(File.join("**", "SIMP", "*.tar.gz"))
+        tar_files.each {|file| File.move file, "#{binaries_dir}/Tarballs"}
+      end
     end
+    # Purge dir
+    FileUtils.rm_r dir
   end
-  # Purge dir
-  FileUtils.rm_r dir
-end
+# Purge ISOS
+  FileUtils.rm_r iso_cache
 end
 # vim: set expandtab ts=2 sw=2:
